requirements.txt
librosa
tqdm
pandas
scikit-learn
mido
pyyaml